ASN mapping for BGP AS reporting.
Adds the ability for clerk to assign and send ASNs with its IPFIX data.  ASNs are currently read from a CSV file, which is by default reread once a day.

To save some time, ASNs aren't added to flows in the packet fast path... since ASNs are deterministically pulled from IPs, we only add the ASNs in during reporting, after we've gathered all flows from threads but before we Send() those flows.

One interesting tidbit:  We put ASNs in the flow::Stats, even though it's actually an identifier (and thus should be in the key).  We do this because of the above... once a flow is in the flow table, its key can't be modified.  And since the IP->ASN mapping is deterministic, just having the IPs in the key is enough to make the ASNs "identity" fields.

I've currently found one good source for free, open ASN databases, 'geolite'.  I've included the geolite_asns.py file to pull ASNs from this location and put them into the format we expect for our ASN CSV file.

Files in this CL:
  - asn_map* - actual logic for reading in CSVs and mapping IP addresses to them
  - geolite_asns.py - pull down free DB to local disk for use
  - clerk.cc - creates ASNMap, uses it before Send()ing the flows
  - send.* - Add ASNs into IPFIX templates
  - flow.* - Add fields for ASNs
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=123457148